+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerExcludeDates.test.jsx
@@ -8,5 +8,4 @@ const DateTimePickerExample = () => (
   />
 );
 
-
 export default DateTimePickerExample;