+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogSuccess.test.jsx
@@ -3,9 +3,9 @@ import NotificationDialogVariant from './NotificationDialogVariant.test';
 import { NotificationDialogVariants } from '../../../NotificationDialog';
 
 const NotificationDialogSuccess = () => (
-  <div>
+  <>
     <NotificationDialogVariant variant={NotificationDialogVariants.SUCCESS} />
-  </div>
+  </>
 );
 
 export default NotificationDialogSuccess;