+++ packages/terra-aggregator/CHANGELOG.md
@@ -4,6 +4,21 @@ Changelog
 Unreleased
 ----------
 
+4.32.0 - (November 7, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+4.31.0 - (October 30, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+4.30.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.29.0 - (October 16, 2019)
 ------------------
 ### Changed