+++ packages/terra-application-menu-layout/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.17.0 - (November 7, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 3.16.0 - (October 16, 2019)
 ------------------
 ### Changed