+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -3,6 +3,23 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Changed icon in menu button and button-style to utility variant
+
+6.18.0 - (November 7, 2019)
+------------------
+### Changed
+* Resolved lint warnings for multiple empty lines
+
+6.17.0 - (October 30, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+6.16.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
 
 6.15.0 - (October 16, 2019)
 ------------------