+++ packages/terra-modal-manager/CHANGELOG.md
@@ -4,6 +4,22 @@ Changelog
 Unreleased
 ----------
 
+6.18.0 - (November 7, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+6.17.0 - (October 30, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+6.16.0 - (October 21, 2019)
+------------------
+### Changed
+* Reenabled tests since form-select updated to fix issue.
+* Updated wdio screenshots
+
 6.15.0 - (October 16, 2019)
 ------------------
 ### Changed