+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -3,6 +3,29 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Changed 'NavigationPromptCheckpoint' to have the same logic as 'notification-dialog' changes
+  * 'message' prop has been deprecated, will resolve to 'startMessage'
+  * Added a 'startMessage' string prop that will display a message after 'title'
+  * Added a 'content' node prop that will display the given node in the Dialog after 'startMessage'
+  * Added an 'endMessage' string prop that will display a message after 'content'
+  * Added 'buttonOrder' prop to determine order of accept and reject buttons
+  * 'emphasizedAction' now accepts a 'none' value and defaults to it.
+
+1.17.0 - (November 7, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+1.16.0 - (October 30, 2019)
+------------------
+### Changed
+* Missing aria-labels added in examples.
+
+1.15.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
 
 1.14.0 - (October 16, 2019)
 ------------------