+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -3,6 +3,24 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* WDIO snapshot updates
+
+5.17.0 - (November 7, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+5.16.0 - (October 30, 2019)
+------------------
+### Changed
+* Duplicate IDs in examples changed.
+* Jest snapshot updates.
+
+5.15.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
 
 5.14.0 - (October 16, 2019)
 ------------------